% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characters.R
\docType{data}
\name{characters}
\alias{characters}
\title{Descriptions of character actions in the Claremont run}
\format{A `tibble` with 4,209 rows and 34 variables: `issue`, the issue
  number, `character`, the character name, and 32 more variables containing
  counts of various actions or names of characters with whom the action was
  done.}
\usage{
characters
}
\description{
`characters` is a `tibble` that contains data on character actions
during the Claremont run by issue.
}
\keyword{datasets}
