% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claremontrun-package.R
\docType{package}
\name{claremontrun-package}
\alias{claremontrun}
\alias{claremontrun-package}
\title{claremontrun: X-Men Data from the Claremont Run Project}
\description{
Data from the Claremont Run project (<http://www.claremontrun.com/>), a collection of data about 
  Chris Claremont's time on the X-Men books (1975-1991). claremontrun includes 7 data sets about Claremont's run: `character_visualization`, 
  `characters`, `comic_bechdel`, `covers`, `issue_collaborators`, `location`, and `xmen_bechdel`.
}
\author{
\strong{Maintainer}: Malcolm Barrett \email{malcolmbarrett@gmail.com} (\href{https://orcid.org/0000-0003-0299-5825}{ORCID})

}
\keyword{*}
\keyword{7}
\keyword{A}
\keyword{Claremont}
\keyword{Data}
\keyword{Package}
\keyword{Project}
\keyword{Run}
\keyword{X-Men}
\keyword{`character_visualization`}
\keyword{`characters`}
\keyword{`comic_bechdel`}
\keyword{`covers`}
\keyword{`issue_collaborators`}
\keyword{`location`}
\keyword{`xmen_bechdel`}
\keyword{claremontrun}
\keyword{claremontrun:}
\keyword{data}
\keyword{for}
\keyword{from}
\keyword{includes}
\keyword{internal}
\keyword{sets:}
\keyword{the}
