% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bechdel.R
\docType{data}
\name{comic_bechdel}
\alias{comic_bechdel}
\title{Whether or not an issue of another comic series met the Bechdel test}
\format{A `tibble` with 308 rows and 9 variables: `series`, the series title,
  `issue`, the issue number, `title`, the title of the issue, `writer`,
  `artist`, `cover_artist`, `pass_bechdel`, whether or not the issue passed
  the Bechdel test, `page_number`, the page number(s) where the issue passed
  the Bechdel test, and `notes`, notes by the Claremont project team
  regarding the issue. done.}
\usage{
comic_bechdel
}
\description{
`comic_bechdel` is a `tibble` that contains data on other (non-X-Men) iconic
runs of Marvel comicbooks, and whether these titles passed the Bechdel test:
1) more than one female character 2) who have a conversation with each other
3) about something other than men. `comic_bechdel`  has data on iconic runs
from . See also [`xmen_bechdel`], which contains Bechdel data on Uncanny
X-Men issues 1-534.
}
\keyword{datasets}
