% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_collaborators.R
\docType{data}
\name{issue_collaborators}
\alias{issue_collaborators}
\title{Data about other collaborators on each issue, such as editors.}
\format{A `tibble` with 190 rows and 4 variables: `issue`, the issue number,
  `editor_in_chief`, `editor`, and `penciller`.}
\usage{
issue_collaborators
}
\description{
`issue_collaborators` is a `tibble` that that contains data on other people
involved with each issue, inclding the editor-in-chief, the editor, and the
penciller.
}
\keyword{datasets}
