% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\docType{data}
\name{locations}
\alias{locations}
\title{Locations of issues in the Claremont run}
\format{A `tibble` with 1,413 rows and 4 variables: `issue`, the issue
  number, `locations`, the location that was shown, `context`, the context of
  the location (e.g. "Present" or "Dream"), and `notes`, which contain notes
  from the Claremont Run project researchers regarding the location.}
\usage{
locations
}
\description{
`locations` is a `tibble` that contains data on the locations shown in
each issue of the Claremont run.
}
\keyword{datasets}
